Added the functionality for re-sizing Projects and Profile images
Script lives on the root of the project, paths need to change if the script is moved

+++ imagereducer.sh
@@ -1,14 +1,21 @@
 #!/bin/bash
 
-profilereduce()
+imagereduce()
 {
 	for file in *;
 	do
-		if [[ ( $file != sm_* ) && ( $file != lg_*) ]]
+		if [[ ( $file != sm_* ) && ( $file != lg_*) && ( $file != *placeholder* ) ]]
 			then
-			if [[ ( $file == *.jpg ) || ( $file == *.png ) || ( $ifle == *.JPG ) ]]
+			if [[ ( $file == *.jpeg ) || ( $ifle == *.JPEG ) || ( $file == *.jpg ) || ( $ifle == *.JPG ) || ( $file == *.png ) || ( $file == '*.PNG' ) ]]
 				then
-				convert "$file" -resize "$LARGE" "lg_$file" && convert "$file" -resize "$SMALL" "sm_$file" && rm "$file"
+				# we are in the profile folder
+				if($flag == "true")
+					then
+					convert "$file" -resize "$LARGE" "lg_$file" && convert "$file" -resize "$SMALL" "sm_$file" && rm "$file"
+				# we are in the project folder
+				else
+					convert "$file" -resize "$LARGE" "lg_$file" && rm "$file"
+				fi
 			fi
 		fi
 	done
@@ -18,9 +25,17 @@ PROFILE_DIRECTORY="public/user-profile/image/"
 PROJECT_DIRECTORY="public/imgs/projects/"
 LARGE="200x175"
 SMALL="150x150"
+flag="true"
+echo "cd into $PROFILE_DIRECTORY"
 cd $PROFILE_DIRECTORY
-profilereduce
-cd ../
+echo "inside $PROFILE_DIRECTORY"
+imagereduce $flag
+echo "finished"
+flag="false"
+cd ../../../
+echo "cd into $PROJECT_DIRECTORY"
 cd $PROJECT_DIRECTORY
-imagereduce
+echo "inside $PROJECT_DIRECTORY"
+imagereduce $flag
+echo "finished"
